move logic into library

+++ strategy-game-backend/src/main/kotlin/de/ruegnerlukas/strategygame/backend/ports/provided/game/GameConnectAction.kt
@@ -8,6 +8,6 @@ interface GameConnectAction {
 	object GameNotFoundError : GameConnectActionError()
 	object InvalidPlayerState : GameConnectActionError()
 
-	suspend fun perform(userId: String, gameId: String, connectionId: Int): Either<GameConnectActionError, Unit>
+	suspend fun perform(userId: String, gameId: String, connectionId: Long): Either<GameConnectActionError, Unit>
 
 }
